[vue] prepare login page
